refactor: AddRequest -> Splliter로 변경
